# Allow unsafe access into MapSector::m_blocks