@lhofhansl 2024-01-09
@sfan5 How about this? Removed non-const access to the new method and changed all callers to use const MapSectors.