@lhofhansl 2024-01-09
Perhaps I should allow the old `getBlocks` from a const object as well.
Not sure about `getBlockNoCreateNoEx` as it modifies the internal cache, I guess it is still semantically read-only.

Both are not needed for this change, but be good for consistency.