@sfan5 2024-01-09
> Not sure about `getBlockNoCreateNoEx` as it modifies the internal cache, I guess it is still semantically read-only.

The correct way to do that is to mark the cache variables are `mutable` so they can still be used on a const object.
But let's not over-complicate this PR.